src/test/java/uk/tw/energy/service/MeterReadingServiceTest.java
package uk.tw.energy.service;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

import java.util.ArrayList;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class MeterReadingServiceTest {

  private MeterReadingService meterReadingService;

  @BeforeEach
  public void setUp() {
    meterReadingService = new MeterReadingService(new ConcurrentHashMap<>());
  }

  @Test
  public void givenMeterIdThatDoesNotExistShouldReturnNull() {
    assertThat(meterReadingService.getReadings("unknown-id")).isEqualTo(Optional.empty());
  }

  @Test
  public void givenMeterReadingThatExistsShouldReturnMeterReadings() {
    meterReadingService.storeReadings("random-id", new ArrayList<>());
    assertThat(meterReadingService.getReadings("random-id"))
        .isEqualTo(Optional.of(new ArrayList<>()));
  }
}
